Implemented libds client call in PlayerExternalsRdkInterface
